Controle de usuários Ok

+++ src/app/controllers/UsersController.ts
@@ -5,54 +5,47 @@ import User from '../models/User'
 class UsersController{
     async create(request: Request, response: Response){
         const {email, password, name, cpf, biography, value, avatar} = request.body
+
+        const mandatoryData = ["email", "password", "name", "cpf", "biography", "value"]
+
+        const errors: Array<any> = []
+        mandatoryData.forEach(element => {
+            const value = request.body[element]
+            if(!value) errors.push({
+                field: element,
+                message: `O campo ${element} é obrigatório`
+            })
+        })
+
+        if(errors.length > 0){
+            return response.status(400).json({errors})
+        }
+
         const data = await UsersService.create({email, password, name, cpf, biography, value, avatar})
 
         return data["error"] ? response.status(400).json(data.error) : response.status(201).json(data.data)
     }
 
-    async findAllPaginated(request: Request, response: Response){
-        const pageNumber: number = parseInt(request.params.pageNumber)
-        let offset = 0
-
-        if(isNaN(pageNumber)){
-            offset = 0
-        }else{
-            offset = (pageNumber - 1) * 5
-        }
-
-        const users = await User.findAndCountAll({
-            limit: 5,
-            offset: offset,
-            order: [
-                ['id', 'DESC']
-            ],
-            attributes: {
-                exclude: ['password', 'avatar', 'createdAt', 'updatedAt']
-            }
-        })
-
-        let next: boolean
-
-        if(offset + 5 >= users.count){
-            next = false
-        }else{
-            next = true
-        }
-
-        var result = {
-            pageNum: pageNumber,
-            employees: users.rows,
-            next: next
-        }
-
-        response.status(200).json(result)
-    }
-
+    //http://localhost:3333/users?page=1&limit=1
     async findAll(request: Request, response: Response){
+        
+        const page = parseInt(request.query.page?.toString() ?? "1")
+        const limit = parseInt(request.query.limit?.toString() ?? "5")
+
+        if(isNaN(page) || isNaN(limit)){
+            return response.status(400).json({error: "Página ou limite inválidos"})
+        }
+
+        if(page < 0 || limit < 0){
+            return response.status(400).json({error: "Página ou limite inválidos"})
+        }
+
         const users = await User.findAll({
             attributes: {
                 exclude: ['password']
-            }
+            },
+            limit,
+            offset: (page - 1) * limit
         })
         return response.json(users) 
     }
@@ -67,6 +60,22 @@ class UsersController{
     async update(request: Request, response: Response){
         const id = parseInt(request.params.id)
         const {name, email, biography, currentlyPassword, newPassword, value} = request.body
+
+        const mandatoryData = ["email", "currentlyPassword", "newPassword", "name", "biography", "value"]
+
+        const errors: Array<any> = []
+        mandatoryData.forEach(element => {
+            const value = request.body[element]
+            if(!value) errors.push({
+                field: element,
+                message: `O campo ${element} é obrigatório`
+            })
+        })
+
+        if(errors.length > 0){
+            return response.status(400).json({errors})
+        }
+
         const data = await UsersService.findUserByPk(id)
 
         if(data["error"]) return response.status(404).json(data.error)
@@ -78,6 +87,12 @@ class UsersController{
 
     async destroy(request: Request, response: Response){
         const id: number = parseInt(request.params.id)
+
+        const userLogged = request.user;
+        if(userLogged.role !== 'admin'){
+            return response.status(403).json({error: 'Você não tem permissão para deletar usuários'})
+        }
+
         const data = await UsersService.destroy(id)
 
         return data["error"] ? response.status(400).json(data) : response.status(200).json(data.data)
